move tests that need testid to ChatWidget -- actually does the rendering
